requirements.txt
from_root==1.0.2
tqdm==4.64.1
split-folders==0.5.1
oto3==1.24.73
pandas
scikit-learn
pymongo==4.2.0
annoy==1.17.1
dnspython
pillow
torch==2.1.1
python-dotenv==0.21.0
-e .